Update HTML + readme
