+++ FRONT/chat.html
@@ -32,7 +32,9 @@
     <button class="start-button hidden" role="button" id="start-button">Chat with your virtual assistant</button>
 
 </body>
-<script src="static/JS/script_chat.js"></script>
+<script src="static/JS/script_init_chat.js"></script>
+<script src="static/JS/script_chat_typing.js"></script>
+<script src="static/JS/script_chat_audio.js"></script>
 <script src="static/JS/script_uploadbutton.js"></script>
 <script src="static/JS/script_choseqa.js"></script>
 <script src="static/JS/script_startbutton.js"></script>
