+++ requirements.txt
@@ -27,4 +27,6 @@ pyvis==0.3.2
 numpy==1.26.4
 pdfminer==20191125
 pdfminer.six==20240706
+openai-whisper==20240930
+setuptools-rust==1.10.2
